review feedback from john
