+++ mg-api-types/versions/src/impls/rdb/mod.rs
@@ -4,4 +4,3 @@
 
 pub(crate) mod constants;
 mod path;
-mod prefix;
